Moving methods in cpp to hpp; placing logger class into local namespace
The tester is, as of this commit, nonfunctional.

In this commit, the Logger class is moved from the 'exceptional'
namespace to the local namespace in exceptional.hpp, and all
method definitions in exceptional.cpp are moved into exceptional.hpp
to allow for this.
The exceptional namespace still exists, and contains the actual
Logger objects used for logging from other files.

This is to allow the logger to be used across multiple files in a
single project more easily and without having to redeclare a global
Logger object in each file.